backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
haystack-ai==2.0.0a2